Fix missing `ngx_udp_connect` compilation issue
Since commit nginx/nginx@0759f088a532ec48170ca03d694cc103757a0f4c the
`ngx_udp_connect` is defined as `static`, hence, prohibiting to use it on an
external module. This patch copy/paste the original function. We will be able to
remove it if/when the `ngx_udp_connect` becomes exported in the headers.